Initial commit for Recorder functionality. (JITaaS)
Recorder is part of the record and replay functionality.
Record and Replay is a feature to enable JIT as a Service (JITaaS).
The ultimate goal for such feature is to remove the JIT from runtimes
And provide it as a service. The recorder part of this feature, as the name implies,
will record the IL at the client side and send it to be compiled in the server.
The server then replays/parses the IL and compiles it, to then send machine
code back to the client so that it can execute the requested IL. The IL contains
services that have recording support (e.g. Add, Sub, Xor, Store, Loop, etc. More to come).
This commit is dedicated to the Recorder functionality. The replay implementation
will be introduced in a later commit.
Equivalent Recorder files for MethodBuilder, IlBuilder, JitBuilder, and
BytecodeBuilder were created. Recorder files became parents of their
respective files. More details as follows.

MethodBuilderRecorder records JitBuilder IL (jbil) for the JitBuilder
services created by MethodBuilder. It also holds a pointer to a
JitBuilderRecorder object and uses that recorder object to record
the services and parameters for all the JitBuilder API called on a
MethodBuilder object, starting from when it is created (using
STATEMENT_NEWMETHODBUILDER) to the calls to, for example,
DefineName or DefineParameter, etc.

IlBuilderRecorder implements every JitBuilder service but only
handles the creation of IlValues and the jbil recording.
IlBuilder retains the code required to actually generate compiler
IL for compiling the target method.

We also added a flag “shouldCompile” in compiler/compile/OMRCompilation.
This flag is used by recordMethodBuilder defined in Jit.hpp/cpp .
When called, this method disables compilation and only records the services
from the callee. Why wouldn’t we want to compile? That’s the ultimate goal of
JITaaS, where the client application only records the services being called (not compile)
then sends these services to the server to be compiled. Once compiled, the
server sends machine code back so that the client application can continue execution.

JitBuilderRecorder defines the basic structure of a JitBuilder IL (ljbil)
representation. The following files provide implementations that
write this representation to a binary buffer, binary file, text file, respectively:
OMRJitBuilderRecorderBinaryBuffer
OMRJitBuilderRecorderBinaryFile
OMRJitBuilderRecorderTextFile

The Replay implementation will subsequently become
children of MethodBuilder, IlBuilder, etc. once completed. e.g.:
IlInjector > IlBuilderRecorder > IlBuilder > MethodBuilderRecorder
MethodBuilderRecorder > MethodBuilder > MethodBuilderReplay

Co-authored-by: Sharon Wang <sharon-wang-cpsc@outlook.com>
Co-authored-by: Mark Stoodley <mstoodle@ca.ibm.com>
Co-authored-by: Andrew Young <youngar17@gmail.com>
Co-authored-by: Charlie Gracie <charlie.gracie@gmail.com>

Signed-off-by: Igor Braga <higorb1@gmail.com>

Update constant names for recorder to be static

Calls such as: rec->STATEMENT_DEFINENAME
Became:        StatementName::STATEMENT_DEFINENAME
Remove imperium library from CMakeLists.

Signed-off-by: Igor Braga <higorb1@gmail.com>

Fix include in StatementNames.

Signed-off-by: Igor Braga <higorb1@gmail.com>

Delete unnecessary VSCode setting.

Signed-off-by: Igor Braga <higorb1@gmail.com>