Initial commit for Replay functionality. (JITaaS)
Replay is part of the record and replay functionality.
Record and Replay is a feature to enable JIT as a Service (JITaaS).
The ultimate goal for Record and Replay is to have the program run as a client,
and have the server “replay” the client’s “recorded” services.
The Replay component of this feature will read IL from a text file,
“rebuild” the IL and compile the services indicated in the file.
Replay compiles services as if the program was calling buildIL(),
without the need to recompile pre-recorded code.
Machine code is sent back to the client so it can continue execution.

Equivalent Replay files for MethodBuilderand JitBuilder were created.
Replay files are children of these files. More details as follows.

MethodBuilderReplay inherits all functionality from its parent MethodBuilder,
which inherits from MethodBuilderRecorder.
MethodBuilderReplay serves as a driver for the replay functionality,
by parsing the IL.

JitBuilderReplay, just like JitBuilderRecorder, defines the basic
structure of a JitBuilder IL (ljbil) representation. The following files
provide implementations that replay this representation to a
binary buffer, binary file and text file, respectively:
OMRJitBuilderReplayBinaryBuffer    (future commit)
OMRJitBuilderReplayBinaryFile      (future commit)
OMRJitBuilderReplayTextFile        (included in this commit)

The Replay files are children of MethodBuilder. For example:
IlInjector > IlBuilderRecorder > IlBuilder > MethodBuilderRecorder
MethodBuilderRecorder > MethodBuilder > MethodBuilderReplay

Co-authored-by: Igor Braga <higorb1@gmail.com>
Co-authored-by: Mark Stoodley <mstoodle@ca.ibm.com>
Co-authored-by: Andrew Young <youngar17@gmail.com>
Co-authored-by: Charlie Gracie <charlie.gracie@gmail.com>

Signed-off-by: Sharon Wang <sharon-wang-cpsc@outlook.com>