Allow systems that have huge pages to use mmap instead of shmget+shmat
Replace shmget + shmat on linux with mmap call when huge pages are used.
Instead of using these calls we use the combination of memfd_create(2)
and mmap(2). For double map to work we need a file descriptor, to get
one we use shm_open if normal page sizes are being used and use that to
mmap. However, if large (huge) pages are being used that would not work
The problem with mmap in linux is that we can only mmap large pages
anonymously. Therefore, if I create a file descriptor using shm_open(3)
and then try to mmap with huge pages with the respective file descriptor
the mmap call fails. It would only succeed if we pass MAP_ANON, but doing
so ignores the file descriptor which is the opposite of what we want. In
a new glibc version (glibc2.27 onwards) there’s a new function that does
exactly what we want, and that's memfd_create(2). This function is very
similar to shm_open(3); however, it returns a file descriptor that has
the same semantics as other anonymous memory allocations. Doing so, we
can mmap(2) huge pages without having to pass MAP_ANON to mmap(2).

Add #if __GLIBC_PREREQ(2,27) to protect against glibc version

Signed-off-by: Igor Braga <higorb1@gmail.com>