Update commit call to populate memory for double map
We need a way to reserve shared memory in case of
double mapping. In Linux systems there's a limit
on the amount of shared memory that can be used at
a given time for a system, that's normally half
the physical memory. Unfortunetely, there's no
straight forward way to reserve shared memory
without commiting it, but we still need a way to
reserve it. Therefore we add a call to mmap with
MAP_POPULATE to reserve/commit memory in case
of a shared heap (double mapping is enabled)
in order to reserve shared memory and prevent
a crash if we reach this limit. MAP_POPULATE
does touches the memory and mincore() checks
if we actually populated the memory. If this
check fails we know that we reached the maximum
shared memory of the system therefore we fail
the commit.

Signed-off-by: Igor Braga <higorb1@gmail.com>